Completely switch to constants
